style: 지도 스크립트 관련 로직 분리
- 지도 컴포넌트 MapLayout으로, 로직은 useSdkLoad 훅으로 분리
